+++ src/impls/$split-when/__tests__/$split-when.test.ts
@@ -4,43 +4,19 @@ import { $splitWhen } from 'iter-tools-es';
 import { $wrap, $unwrapDeep } from '../../../test/$helpers.js';
 
 describe($`splitWhen`, () => {
-  describe('when there are no values', () => {
-    it(
-      'yields empty parts',
-      $async(() => {
-        const [first, second] = $splitWhen((_, i) => i > 0, $wrap([]));
-        expect($await($unwrapDeep([first, second]))).toEqual([[], []]);
-      }),
-    );
-  });
+  it(
+    'should split between every value which is equal to the on argument',
+    $async(() => {
+      expect(
+        $await($unwrapDeep($splitWhen((i) => i === null, $wrap([1, null, 2, null, 3])))),
+      ).toEqual([[1], [2], [3]]);
+    }),
+  );
 
-  describe('when `predicate(value, i)` goes from falsy to truthy', () => {
-    it(
-      'puts values in each part',
-      $async(() => {
-        const [first, second] = $splitWhen((_, i) => i > 0, $wrap([1, 2, 3]));
-        expect($await($unwrapDeep([first, second]))).toEqual([[1], [2, 3]]);
-      }),
-    );
-  });
-
-  describe('when `predicate(value, i)` is truthy initially', () => {
-    it(
-      'puts all the values in the second part',
-      $async(() => {
-        const [first, second] = $splitWhen((v) => v > 0, $wrap([1, 2, 3]));
-        expect($await($unwrapDeep([first, second]))).toEqual([[], [1, 2, 3]]);
-      }),
-    );
-  });
-
-  describe('when `predicate(value, i)` is never truthy', () => {
-    it(
-      'puts all the values in the first part',
-      $async(() => {
-        const [first, second] = $splitWhen((_) => null, $wrap([1, 2, 3]));
-        expect($await($unwrapDeep([first, second]))).toEqual([[1, 2, 3], []]);
-      }),
-    );
-  });
+  it(
+    'should return no parts if source is empty',
+    $async(() => {
+      expect($await($unwrapDeep($splitWhen((i) => i, null)))).toEqual([]);
+    }),
+  );
 });